Add files via upload
This 'QUIZ GAME' Project is designed for a question in which you can generate and manage simple database for questions. The question number is automatically generated by the software and is stored in a binary file by the name 'QUESTION'. This data base is basically used as a MASTER file to be used as a look-up table for information like Quiz Option, of C language. In this software you can ask for the question of c language. If he is incorrect then he lost the game & game is closed